requirements.txt
# WSR Generator - Requirements
# Python 3.8+

# Base dependencies
pandas>=1.5.0
numpy>=1.23.0
psycopg2-binary>=2.9.0
python-dotenv>=0.20.0

# Optional for LLM integration (if needed)
openai>=1.0.0
requests>=2.28.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Logging and monitoring
colorlog>=6.6.0

# Data validation
pydantic>=1.10.0

# Proyección Objetiva (motor estadístico Holt-Winters)
statsmodels>=0.14.0
scipy>=1.10.0